Separate StakeUnstake Controller
