+++ shhhtoshi-backend/ShhhToshiApp/DTOs/WalletInfoDto.cs
@@ -6,6 +6,7 @@
         public decimal TONBalance { get; set; }
         public decimal StakedAmount { get; set; }
         public DateTime LastStakedAt { get; set; }
+        public DateTime LastUnstakedAt { get; set; }
     }
 
 }
